requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
python-decouple==3.8
psycopg2-binary==2.9.6
django-cors-headers==4.0.0
Pillow==9.5.0
